+++ lib/mavlink/serial_connection.ex
@@ -3,15 +3,14 @@ defmodule XMAVLink.SerialConnection do
   XMAVLink.Router delegate for Serial connections
   """
 
-  @smallest_mavlink_message 8
+  @behaviour XMAVLink.Transport
 
   require Logger
 
+  alias XMAVLink.Connection.Inbound
   alias XMAVLink.ConnectionWorker
   alias XMAVLink.Frame
   alias Circuits.UART
-
-  import XMAVLink.Frame, only: [binary_to_frame_and_tail: 1, validate_and_unpack: 3]
 
   defstruct port: nil,
             baud: nil,
@@ -34,47 +33,14 @@ defmodule XMAVLink.SerialConnection do
         receiving_connection = %XMAVLink.SerialConnection{buffer: buffer},
         dialect
       ) do
-    case binary_to_frame_and_tail(buffer <> raw) do
-      :not_a_frame ->
-        # Noise or malformed frame
-        if byte_size(buffer) + byte_size(raw) > 0 do
-          :ok =
-            Logger.debug("SerialConnection.handle_info: Not a frame: #{inspect(buffer <> raw)}")
-        end
-
-        {:error, :not_a_frame, port, struct(receiving_connection, buffer: <<>>)}
-
-      {nil, rest} ->
-        {:error, :incomplete_frame, port, struct(receiving_connection, buffer: rest)}
-
-      {received_frame, rest} ->
-        # Rest could include a complete message, return later to try emptying the buffer
-        if byte_size(rest) >= @smallest_mavlink_message,
-          do: send(self(), {:circuits_uart, port, <<>>})
-
-        connection = struct(receiving_connection, buffer: rest)
-
-        case validate_and_unpack(received_frame, dialect, receiving_connection.signing) do
-          {:ok, valid_frame, signing} ->
-            {:ok, port, struct(connection, signing: signing), valid_frame}
-
-          {:unknown_message, signing} ->
-            # We re-broadcast valid frames with unknown messages
-            :ok =
-              Logger.debug("rebroadcasting unknown message with id #{received_frame.message_id}")
-
-            {:ok, port, struct(connection, signing: signing),
-             struct(received_frame, target: :broadcast)}
-
-          {:error, reason, signing} ->
-            :ok =
-              Logger.debug(
-                "SerialConnection.handle_info: frame received failed: #{Atom.to_string(reason)}"
-              )
-
-            {:error, reason, port, struct(connection, signing: signing)}
-        end
-    end
+    Inbound.stream(
+      raw,
+      receiving_connection,
+      buffer,
+      port,
+      dialect,
+      "SerialConnection.handle_info"
+    )
   end
 
   def open(["serial", port, baud], controlling_process) do