+++ lib/mavlink_util/defaults.ex
@@ -1,0 +1,7 @@
+defmodule XMAVLink.Util.Defaults do
+  @moduledoc false
+
+  @default_dialect :"Elixir.Common"
+
+  def default_dialect, do: @default_dialect
+end