+++ lib/mavlink_util/supervisor.ex
@@ -2,6 +2,7 @@ defmodule XMAVLink.Util.Supervisor do
   @moduledoc false
 
   use Supervisor
+  alias XMAVLink.Util.Context
 
   def start_link(opts) do
     Supervisor.start_link(__MODULE__, opts, name: __MODULE__)
@@ -9,12 +10,16 @@ defmodule XMAVLink.Util.Supervisor do
 
   @impl true
   def init(opts) do
-    router = Keyword.get(opts, :router, XMAVLink.Router)
+    context = Context.new(opts)
     auto_param_request = Keyword.get(opts, :auto_param_request, true)
 
     children = [
-      {XMAVLink.Util.FocusManager, %{}},
-      {XMAVLink.Util.CacheManager, %{router: router, auto_param_request: auto_param_request}}
+      {XMAVLink.Util.FocusManager, %{context: context}},
+      {XMAVLink.Util.CacheManager,
+       %{
+         context: context,
+         auto_param_request: auto_param_request
+       }}
     ]
 
     Supervisor.init(children, strategy: :one_for_one)